updt

+++ src/pages/Profile.jsx
@@ -3,7 +3,9 @@ import FormSemiHeader from "../components/FormSemiHeader"
 import axios from "axios"
 import { useNavigate } from "react-router-dom"
 import UpdateModal from "../components/updateModal"
-
+import { TbPasswordUser } from "react-icons/tb";
+import { MdDelete } from "react-icons/md";
+import { FaPenToSquare } from "react-icons/fa6";
 
 
 function Profile() {
@@ -29,8 +31,12 @@ function Profile() {
  }
 
  const deleteUser = async()=>{
-    try {
+    const confirmDelete = window.confirm("Are you sure you want to delete your account ? ")
+   
+    if(confirmDelete){
+         try {
       const token = localStorage.getItem("token")
+
       await axios.delete("http://localhost:3000/api/v1/user/deleteuser",{
         headers : {
           authorization : `Bearer ${token}`
@@ -42,7 +48,9 @@ function Profile() {
     } catch (error) {
       console.log("error : ",error)
     }
+    }
  }
+
 
 
  useEffect(()=>{
@@ -54,39 +62,40 @@ function Profile() {
  },[user])
 
   return (
-    <div className='dark:bg-black/90 dark:text-white min-h-screen flex flex-col justify-center items-center'>
-      <h1>User Profile</h1>
+    <div className='dark:bg-black/90 dark:text-white min-h-screen flex flex-col justify-center items-center px-2'>
+     
       
-      <div>
-        <div className="flex flex-col gap-y-6">
+      <div className="border dark:border-white/40 px-3 py-6 rounded-md">
+         <h1 className=" text-xl text-center  mb-8 rounded-full">User Profile</h1>
+        <div className="flex flex-col gap-y-8 ">
             
-            <div className="flex flex-col gap-x-4 mx-auto">
+            <div className="flex flex-col gap-x-4 mx-auto ">
                 <FormSemiHeader label={"Username"} />
                 <div className="flex ">
-                      <input placeholder={user?.username || "Loading..."} readOnly={true} type="text" className="border rounded-md p-2 min-w-xs mr-3" />
+                      <input placeholder={user?.username || "Loading..."} readOnly={true} type="text" className="border rounded-md p-2 min-w-2xs " />
                 </div>
             </div>
 
             <div className="flex flex-col gap-x-4 mx-auto">
                 <FormSemiHeader label={"email"} />
                 <div className="flex ">
-                  <input placeholder={user?.email || "Loading..."} readOnly={true} type="text" className="border rounded-md p-2 min-w-xs mr-3" />
+                  <input placeholder={user?.email || "Loading..."} readOnly={true} type="text" className="border rounded-md p-2 min-w-2xs " />
                 </div>
             </div>
 
-            <div className="mx-auto space-x-2">
-                <button className="w-[150px] py-2 border dark:hover:bg-gray-600 transition-transform duration-200 cursor-pointer">
-                  Change Password
+            <div className="mx-auto space-x-4">
+                <button className=" px-3 py-2 border dark:hover:bg-gray-600 transition-transform duration-200 cursor-pointer">
+                  <TbPasswordUser size={20}/>
                 </button>
-                <button onClick={deleteUser}  className="w-[150px] py-2 border dark:hover:bg-gray-600 transition-transform duration-200 cursor-pointer">
-                  Delete User
+                <button onClick={deleteUser}  className="px-3 py-2 border dark:hover:bg-gray-600 transition-transform duration-200 cursor-pointer">
+                  <MdDelete size={20}/>
                 </button>
-                <button onClick={()=>setModal(true)} className="w-[150px] py-2 border dark:hover:bg-gray-600 transition-transform duration-200 cursor-pointer">
-                  Update
+                <button onClick={()=>setModal(true)} className="px-3 py-2 border dark:hover:bg-gray-600 transition-transform duration-200 cursor-pointer">
+                  <FaPenToSquare size={20}/>
                 </button>
             </div>
 
-             <UpdateModal modal={modal} setModal={setModal}/>
+             <UpdateModal refatch = {getUser} modal={modal} setModal={setModal}/>
              
             
         </div>
